requirements.txt
sounddevice
soundfile
pyyaml
numpy
